[REFACTOR] refactor recipeList api

+++ backend/src/main/java/com/ssafy/letcipe/api/service/RecipeListService.java
@@ -15,8 +15,9 @@ import com.ssafy.letcipe.domain.recipeListBookmark.RecipeListBookmarkRepository;
 import com.ssafy.letcipe.domain.recipeListItem.RecipeListItem;
 import com.ssafy.letcipe.domain.recipeListItem.RecipeListItemRepository;
 import com.ssafy.letcipe.domain.user.User;
-import com.ssafy.letcipe.domain.user.UserRepository;
+import com.ssafy.letcipe.exception.AuthorityViolationException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -27,10 +28,10 @@ import java.util.List;
 @RequiredArgsConstructor
 public class RecipeListService {
     private final RecipeListRepository recipeListRepository;
-    private final UserRepository userRepository;
     private final RecipeListBookmarkRepository recipeListBookmarkRepository;
     private final RecipeRepository recipeRepository;
     private final RecipeListItemRepository recipeListItemRepository;
+    private final UserService userService;
 
     public RecipeList findRecipeList(Long recipeListId) {
         return recipeListRepository.findById(recipeListId).orElseThrow(() -> new NullPointerException("레시피 리스트를 찾을 수 없습니다."));
@@ -42,8 +43,8 @@ public class RecipeListService {
         return new ResGetRecipeListDto(recipeList);
     }
 
-    public void createRecipeList(String accessToken, ReqCreateRecipeListDto reqCreateRecipeListDto) {
-        User user = userRepository.findById(1L).orElseThrow();
+    public void createRecipeList(Long userId, ReqCreateRecipeListDto reqCreateRecipeListDto) {
+        User user = userService.findUser(userId);
         RecipeList recipeList = RecipeList.builder()
                 .user(user)
                 .name(reqCreateRecipeListDto.getName())
@@ -53,23 +54,24 @@ public class RecipeListService {
         recipeListRepository.save(recipeList);
     }
 
-    public void updateRecipeList(String accessToken, ReqUpdateRecipeListDto reqUpdateRecipeListDto, Long recipeListId) {
-        User user = userRepository.findById(1L).orElseThrow();
-        RecipeList recipeList = recipeListRepository.findById(recipeListId).orElseThrow();
+    public void updateRecipeList(Long userId, ReqUpdateRecipeListDto reqUpdateRecipeListDto, Long recipeListId) {
+        RecipeList recipeList = findRecipeList(recipeListId);
+        if (!userId.equals(recipeList.getUser().getId())) throw new AuthorityViolationException("수정 권한이 없습니다.");
         recipeList.updateRecipeList(reqUpdateRecipeListDto.getName(), reqUpdateRecipeListDto.getDescription(), reqUpdateRecipeListDto.getIsShared());
         recipeListRepository.save(recipeList);
     }
 
-    public void deleteRecipeList(String accessToken, Long recipeListId) {
-        User user = userRepository.findById(1L).orElseThrow();
-        RecipeList recipeList = recipeListRepository.findById(recipeListId).orElseThrow();
+    public void deleteRecipeList(Long userId, Long recipeListId) {
+        RecipeList recipeList = findRecipeList(recipeListId);
+        if (!userId.equals(recipeList.getUser().getId())) throw new AuthorityViolationException("삭제 권한이 없습니다.");
         recipeList.deleteRecipeList();
         recipeListRepository.save(recipeList);
     }
 
-    public void createRecipeListBookmark(String accessToken, Long recipeListId) {
-        User user = userRepository.findById(1L).orElseThrow();
-        RecipeList recipeList = recipeListRepository.findById(recipeListId).orElseThrow();
+    public void createRecipeListBookmark(Long userId, Long recipeListId) {
+        User user = userService.findUser(userId);
+        RecipeList recipeList = findRecipeList(recipeListId);
+        if (!userId.equals(recipeList.getUser().getId())) throw new AuthorityViolationException("접근 권한이 없습니다.");
         RecipeListBookmark recipeListBookmark = RecipeListBookmark.builder()
                 .user(user)
                 .recipeList(recipeList)
@@ -77,15 +79,15 @@ public class RecipeListService {
         recipeListBookmarkRepository.save(recipeListBookmark);
     }
 
-    public void deleteRecipeListBookmark(String accessToken, Long recipeListId) {
-        User user = userRepository.findById(1L).orElseThrow();
-        RecipeListBookmark recipeListBookmark = recipeListBookmarkRepository.findByUserIdAndRecipeListId(1L, recipeListId).orElseThrow();
+    public void deleteRecipeListBookmark(Long userId, Long recipeListId) {
+        RecipeListBookmark recipeListBookmark = recipeListBookmarkRepository.findByUserIdAndRecipeListId(userId, recipeListId).orElseThrow();
+        if (!userId.equals(recipeListBookmark.getUser().getId())) throw new AuthorityViolationException("접근 권한이 없습니다.");
         recipeListBookmarkRepository.delete(recipeListBookmark);
     }
 
-    public void addRecipeListItem(String accessToken, ReqPostRecipeListItemDto reqPostRecipeListItemDto) {
-        User user = userRepository.findById(1L).orElseThrow();
-        RecipeList recipeList = recipeListRepository.findById(reqPostRecipeListItemDto.getRecipeListId()).orElseThrow();
+    public void addRecipeListItem(Long userId, ReqPostRecipeListItemDto reqPostRecipeListItemDto) {
+        RecipeList recipeList = findRecipeList(reqPostRecipeListItemDto.getRecipeListId());
+        if (!userId.equals(recipeList.getUser().getId())) throw new AuthorityViolationException("접근 권한이 없습니다.");
         Recipe recipe = recipeRepository.findById(reqPostRecipeListItemDto.getRecipeId()).orElseThrow();
         RecipeListItem recipeListItem = RecipeListItem.builder()
                 .recipeList(recipeList)
@@ -94,15 +96,16 @@ public class RecipeListService {
         recipeListItemRepository.save(recipeListItem);
     }
 
-    public void deleteRecipeListItem(String accessToken, ReqDeleteRecipeListItemDto reqDeleteRecipeListItemDto) {
-        User user = userRepository.findById(1L).orElseThrow();
+    public void deleteRecipeListItem(Long userId, ReqDeleteRecipeListItemDto reqDeleteRecipeListItemDto) {
         RecipeListItem recipeListItem = recipeListItemRepository.findByRecipeListIdAndRecipeId(reqDeleteRecipeListItemDto.getRecipeListId(), reqDeleteRecipeListItemDto.getRecipeId());
+        if (!userId.equals(recipeListItem.getRecipeList().getUser().getId()))
+            throw new AuthorityViolationException("접근 권한이 없습니다.");
         recipeListItemRepository.delete(recipeListItem);
     }
 
     @Transactional
-    public List<ResSearchRecipeListDto> searchRecipeList(String keyword) {
-        List<RecipeList> searched = recipeListRepository.findByNameContaining(keyword);
+    public List<ResSearchRecipeListDto> searchRecipeList(Pageable pageable, String keyword) {
+        List<RecipeList> searched = recipeListRepository.findByNameContaining(pageable, keyword);
         List<ResSearchRecipeListDto> result = new ArrayList<>();
         for (RecipeList recipeList : searched) {
             result.add(new ResSearchRecipeListDto(recipeList));
